+++ CHANGELOG.md
@@ -22,10 +22,13 @@ Added
 -----
 - Added tests for the Fredkin gate (#357)
 - Added tests for the cu1 gate (#360)
+- Added tests for statevector and stabilizer snapshots (\#355)
 
 Changed
 -------
 - New (efficient) implementation for expectation_value_pauli (\#344)
+- Stabilizer snapshot returns stabilizer instead of full Clifford table (\#355)
+- Signature of SnapshotStatevector and SnapshotStabilizer (\#355)
 - Changed all names from tensor_network_state to matrix_product_state (\#356)
 - Update device noise model to consume asymmetric readout errors from backends (\#354)
 - Update device noise model to use gate_length (\#352)